pyproject.toml
[tool.poetry]
name = "reformers-model-api-server"
version = "0.1.0"
description = ""
authors = ["Edmund Widl <edmund.widl@ait.ac.at>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = ">=2.6.0"
setuptools = ">=21.0.0"
click = "^8.1.8"
connexion = {version = "~=2.14.2", extras = ["swagger-ui"]}
flask = "2.1.1"
swagger-ui-bundle = "~=0.0.9"

[tool.poetry.scripts]
reformers_model_api_server = "reformers_model_api_server.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
